api/requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
psycopg2>=2.8.6,<2.9
drf-spectacular>=0.15.1,<0.16
django-cors-headers>=3.11.0,<3.13
pillow
cryptography
django-rest-knox
celery
django-celery-results
django-celery-beat
drf-extra-fields
mongoengine
pymongo>=3.6.0
sphinx
